+++ src/ApiManagement/ApiManagement/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Updated description of ResourceId param 'New-AzApiManagementBackend' and 'Set-AzApiManagementBackend' cmdlet [#16868]
 * Fixed Path mandatory bug 'Import-AzureApiManagementApi' cmdlet [#17991]
+* Added parameter `Examples` to `Get-AzApiManagementOperation` response
 
 ## Version 4.0.1
 * Upgraded AutoMapper to Microsoft.Azure.PowerShell.AutoMapper 6.2.2 with fix [#18721]